feat: init configs file and move simple_cli_client configs
